+++ src/main/java/com/maan/eway/bean/PolicyCoverData.java
@@ -20,6 +20,7 @@ import javax.persistence.Table;
 import lombok.*;
 import javax.persistence.*;
 import java.io.Serializable;
+import java.math.BigDecimal;
 import java.util.List;
 
 import java.util.Date;
@@ -122,46 +123,46 @@ private static final long serialVersionUID = 1L;
     private String     calcType ;
 
     @Column(name="MINIMUM_PREMIUM")
-    private Double     minimumPremium ;
+    private BigDecimal     minimumPremium ;
 
     @Column(name="SUM_INSURED")
-    private Double     sumInsured ;
+    private BigDecimal     sumInsured ;
 
     @Column(name="RATE")
-    private Double     rate ;
+    private BigDecimal     rate ;
 
     @Column(name="CURRENCY", length=20)
     private String     currency ;
 
     @Column(name="EXCHANGE_RATE")
-    private Double     exchangeRate ;
+    private BigDecimal     exchangeRate ;
 
     @Column(name="PREMIUM_BEFORE_DISCOUNT_FC")
-    private Double     premiumBeforeDiscountFc ;
+    private BigDecimal     premiumBeforeDiscountFc ;
 
     @Column(name="PREMIUM_BEFORE_DISCOUNT_LC")
-    private Double     premiumBeforeDiscountLc ;
+    private BigDecimal     premiumBeforeDiscountLc ;
 
     @Column(name="PREMIUM_AFTER_DISCOUNT_FC")
-    private Double     premiumAfterDiscountFc ;
+    private BigDecimal     premiumAfterDiscountFc ;
 
     @Column(name="PREMIUM_AFTER_DISCOUNT_LC")
-    private Double     premiumAfterDiscountLc ;
+    private BigDecimal     premiumAfterDiscountLc ;
 
     @Column(name="PREMIUM_EXCLUDED_TAX_FC")
-    private Double     premiumExcludedTaxFc ;
+    private BigDecimal     premiumExcludedTaxFc ;
 
     @Column(name="PREMIUM_EXCLUDED_TAX_LC")
-    private Double     premiumExcludedTaxLc ;
+    private BigDecimal     premiumExcludedTaxLc ;
 
     @Column(name="PREMIUM_INCLUDED_TAX_FC")
-    private Double     premiumIncludedTaxFc ;
+    private BigDecimal     premiumIncludedTaxFc ;
 
     @Column(name="PREMIUM_INCLUDED_TAX_LC")
-    private Double     premiumIncludedTaxLc ;
+    private BigDecimal     premiumIncludedTaxLc ;
 
     @Column(name="FACTOR_TYPE_ID")
-    private Double     factorTypeId ;
+    private BigDecimal     factorTypeId ;
 
     @Column(name="DEPENDENT_COVER_YN", length=20)
     private String     dependentCoverYn ;
@@ -186,10 +187,10 @@ private static final long serialVersionUID = 1L;
     private String     createdBy ;
 
     @Column(name="TAX_RATE")
-    private Double     taxRate ;
+    private BigDecimal     taxRate ;
 
     @Column(name="TAX_AMOUNT")
-    private Double     taxAmount ;
+    private BigDecimal     taxAmount ;
 
     @Column(name="TAX_DESC", length=100)
     private String     taxDesc ;
@@ -207,7 +208,7 @@ private static final long serialVersionUID = 1L;
     private String     taxExemptCode ;
 
     @Column(name="MAX_LODING_AMOUNT")
-    private Double     maxLodingAmount ;
+    private BigDecimal     maxLodingAmount ;
 
     @Column(name="Is_REFERRAL", length=5)
     private String     isReferral ;
